upload_files_drf/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.1
django-celery-results==2.5.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.4
openapi-codec==1.3.2
packaging==23.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
